add meson build support
Meson build support is useful as it simplifies vendoring of
dependencies. Wrap files make it easy to integrate libubootenv and its
dependencies with other projects

This is almost a 1-to-1 port of the CMakeLists. The pkg-config file is
no longer created from a config however, but is generated by Meson's
internal pkg-config library.

Signed-off-by: Brandon Maier <brandon.maier@collins.com>